misc: move mysql query editor into seperate folder
