+++ src/components/QueryEditor/mysql/QueryEditor.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { QueryEditorProps } from '@grafana/data';
-import { DataSource } from '../../datasource';
-import { GreptimeSourceOptions, GreptimeQuery } from '../../types';
+import { DataSource } from 'datasource';
+import { GreptimeSourceOptions, GreptimeQuery } from 'types';
 import { RawQueryEditor } from './RawQueryEditor';
 import { VisualQueryEditor } from './VisualQueryEditor/';
 import { QueryEditorModeSwitcher } from './QueryModeSwitcher';
