+++ src/components/QueryEditor/mysql/VisualQueryEditor/AddSegment.tsx
