+++ src/components/QueryEditor/mysql/VisualQueryEditor/README.md
