+++ src/components/QueryEditor/mysql/VisualQueryEditor/RemoveablePopover.tsx
