+++ src/components/QueryEditor/mysql/VisualQueryEditor/SelectAddons.tsx
