+++ src/components/QueryEditor/mysql/index.tsx
@@ -1,0 +1,1 @@
+export { QueryEditor } from './QueryEditor';
