+++ src/types.ts
@@ -1,7 +1,7 @@
 import type { DataQuery, DataSourceJsonData } from '@grafana/data';
 import { HealthCheckResultDetails } from '@grafana/runtime';
-import type { SelectStatement } from 'components/QueryEditor/VisualQueryEditor/SelectSection';
-import type { WhereStatement } from 'components/QueryEditor/VisualQueryEditor/WhereSection';
+import type { SelectStatement } from 'components/QueryEditor/mysql/VisualQueryEditor/SelectSection';
+import type { WhereStatement } from 'components/QueryEditor/mysql/VisualQueryEditor/WhereSection';
 
 export interface GreptimeQuery extends DataQuery {
   isRawQuery?: boolean;
